# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CINTEXA CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: getSites
      tags: [sites]
      summary: List all sites
      responses:
        "200":
          description: Site list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a new site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get a site by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Pages ───────────────────────────────────────────────────────────────────
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  # ─── Media ───────────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Categories ──────────────────────────────────────────────────────────────
...
  /seo/settings:
    get:
      operationId: getSeoSettings
      tags: [seo]
      summary: Get global SEO settings
      responses:
        "200":
          description: SEO settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoSettings"
    patch:
      operationId: updateSeoSettings
      tags: [seo]
      summary: Update global SEO settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeoSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoSettings"
  /seo/redirects:
    get:
      operationId: getSeoRedirects
      tags: [seo]
      summary: List all redirects
      responses:
        "200":
          description: Redirects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Redirect"
    post:
      operationId: createRedirect
      tags: [seo]
      summary: Create a redirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedirectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redirect"
  /seo/redirects/{id}:
    delete:
      operationId: deleteRedirect
      tags: [seo]
      summary: Delete a redirect
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AI ──────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
        description:
          type: ["string", "null"]
        favicon:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        status:
          type: string
        primaryColor:
          type: ["string", "null"]
        language:
          type: string
        timezone:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, domain, status, language, timezone, createdAt, updatedAt]
    SiteInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
        description:
          type: string
        favicon:
          type: string
        logo:
          type: string
        primaryColor:
          type: string
        language:
          type: string
        timezone:
          type: string
      required: [name, domain]
    SiteUpdate:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        description:
          type: string
        favicon:
          type: string
        logo:
          type: string
        status:
          type: string
        primaryColor:
          type: string
        language:
          type: string
        timezone:
          type: string

    # ─── Page ──────────────────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        status:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        featuredImage:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        readingTime:
          type: ["integer", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, status, createdAt, updatedAt]
    PostInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        authorId:
          type: integer
        categoryId:
          type: integer
        featuredImage:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
      required: [title, slug]
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        status:
          type: string
        excerpt:
          type: string
        content:
          type: string
        authorId:
          type: integer
        categoryId:
          type: integer
        featuredImage:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string

    # ─── Media ─────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required: [id, name, email, role, status, createdAt]
    UserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
        avatar:
          type: string
        bio:
          type: string
      required: [name, email, role]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatar:
          type: string
        bio:
          type: string
        status:
          type: string

    # ─── Category ──────────────────────────────────────────────────────────────
...
    SeoSettings:
      type: object
      properties:
        id:
          type: integer
        siteTitle:
          type: string
        siteDescription:
          type: string
        robots:
          type: string
        googleAnalyticsId:
          type: ["string", "null"]
        googleSearchConsoleId:
          type: ["string", "null"]
        ogImage:
          type: ["string", "null"]
        twitterHandle:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, siteTitle, siteDescription, robots, updatedAt]
    SeoSettingsUpdate:
      type: object
      properties:
        siteTitle:
          type: string
        siteDescription:
          type: string
        robots:
          type: string
        googleAnalyticsId:
          type: string
        googleSearchConsoleId:
          type: string
        ogImage:
          type: string
        twitterHandle:
          type: string
    Redirect:
      type: object
      properties:
        id:
          type: integer
        from:
          type: string
        to:
          type: string
        type:
          type: integer
        createdAt:
          type: string
      required: [id, from, to, type, createdAt]
    RedirectInput:
      type: object
      properties:
        from:
          type: string
          minLength: 1
        to:
          type: string
          minLength: 1
        type:
          type: integer
      required: [from, to, type]

    # ─── AI ────────────────────────────────────────────────────────────────────
...