# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CINTEXA CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  title?: string;
  slug?: string;
  status?: string;
  excerpt?: string;
  content?: string;
  authorId?: number;
  categoryId?: number;
  featuredImage?: string;
  metaTitle?: string;
  metaDescription?: string;
}