# lib/api-zod/src/generated/types/redirect.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CINTEXA CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Redirect {
  id: number;
  from: string;
  to: string;
  type: number;
  createdAt: string;
}