# lib/api-zod/src/generated/types/siteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CINTEXA CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteInput {
  /** @minLength 1 */
  name: string;
  domain: string;
  description?: string;
  favicon?: string;
  logo?: string;
  primaryColor?: string;
  language?: string;
  timezone?: string;
}