# lib/api-zod/src/generated/types/siteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CINTEXA CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteUpdate {
  name?: string;
  domain?: string;
  description?: string;
  favicon?: string;
  logo?: string;
  status?: string;
  primaryColor?: string;
  language?: string;
  timezone?: string;
}